dom: move the importer and mapper to DomEnvironment
Load files and their dependencies by reusing the same importer and
mapper. This can be done by adding the importer and the mapper as
members of the DomEnvironment, now that populateQmlFile (ex-createDom())
is a member of the DomEnvironment.

Also adapt the QQmlJSDomAstCreator to accept a QQmlJSScope in its
constructor instead of creating its own. Like this, all semantic scopes
are created via the same QQmlJSImporter.

QEXPECT_FAIL some tests that stopped working because of the
restructuring, and unexpect_fail them later once the lazy loading is
completely done. The cyclic dependency here is following: if you reuse
the same importer for all qml files, then dependencies need to be loaded
lazily via the importer's importvisitor. If the importer's importvisitor
can load files lazily, then it needs to give the QmlFile a handle to
populate itself. The handle is the qqmljsscope, which is actually
obtained by the importer who needs to be reused for all loaded qml
files. Therefore QEXPECT_FAIL the files until all the component of that
cyclic dependency are done.

Task-number: QTBUG-122645
Change-Id: Idfb7d33d3ccb7c0fcc10c109a8815cbb8e3a3f6b
Reviewed-by: Dmitrii Akshintsev <dmitrii.akshintsev@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>