tst_qquickmenubar: assign the menuBar to 'menuBar', and not 'header'
The MenuBar in the QML test files is assigned to
ApplicationWindow.header instead of
ApplicationWindow.menuBar. This is most likely due
to ApplicationWindow.menuBar not being available
until Qt 5.10.

This patch will change this so that we use
ApplicationWindow.menuBar instead. That way we
actually test the recommended way of assigning a
menubar to ApplicationWindow.

To ensure that we stay backwards compatible, add
a QML file and test case that checks that it
works to use ApplicationWindow.header as well.

Change-Id: I9d4810a1a5c086efb74d5cecc6e44f6ca3ff42d6
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>