Fix baseline offset with RichText when multiple fonts are used
When combining multiple different fonts in one text, we would
align the baseline based on the ascent of the main font when
using Text.RichText. When using StyledText, we would use the
QTextLine's actual ascent instead (the maximum of all fonts
in the line).

This caused a misalignment when there are fonts with larger
ascents in the text (e.g. fallback fonts).

If possible, we use the first block's first line for alignment,
same as for StyledText. Otherwise, we fall back to using the
QFontMetrics like before.

Fixes: QTBUG-122027
Change-Id: I86d12bfcd58cee837caff4ea72e002d4d5723b3b
Reviewed-by: Lars Knoll <lars@knoll.priv.no>