V4: Fix setting of regexp flags
We of course have to amend the previously set flags, not the origin.
This only worked sort of because most of the flags are the same in input
and output.

This causes one test in the ECMAScript suite to pass and three to fail.
Those three were passing for the wrong reason, relying on the confusion
between the unicode and sticky flags. The failures are actually due to a
defect in Yarr that will have to be addressed separately.

Fixes: QTBUG-121509
Change-Id: Id930bba49e12dede2eb699e446eb528072f092b4
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>