QQuickMenu: search for menubar using parentItem
The Menu (and the MenuBarItem) will be reparented into
the MenuBar using parentItem. And there is currently no
requirement that the QObject parent needs to do the
same. So rather than meddling with the QObject parent
for the Menu and the MenuBarItem, simply search for
the MenuBar using parentItem instead. This is also more
aligned with the QML Item.parent property, which
follows the parentItem hierarchy, and not the
parent hierarchy.

Change-Id: I112b4aa590e05505d63e090d392672e89d7c2fac
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>