dom: enable populateQmlFile as visitor in importer
Change qqmljsimporter to accept a custom importvisitor instead of a
custom importvisitorcreator. This gives more flexibility, for example to
allow to use populateQmlFile() in combination with qqmljsimporter.

Extend DomEnvironment::ensureImporter() to set up a custom
visitor in the importer that calls populateQmlFile() to construct the
Dom lazily.

Adapt the qmltc custom visitor creator to be a custom visitor.

Add a basic test to see if the qqmljsscope created on two files with the
populateQmlFile() custom visitor share the same address. This is a
requirement for the find usages functionality.

Also run the importvisitor even when parsing fails:
The importvisitor for the Dom will parse the code again with parser
recovery, and therefore the importvisitor has to always be run to get
the parser recovery working in the tests.

Do not try to load dependencies in LoadInfo::doAddDependencies, because
it will trigger QmlFile lazyloading. This happens because dependencies
in the form of import statements can only be processed if
qqmldomastcreator already collected all the import statements.

Instead, load the dependencies in qqmldomastcreator (that is, during the
lazy loading) when encountering import statements. Like this, a lazy QML
file will only load its imported modules when its getting populated.
Also make sure to not load anything if NoDependencies is enabled in the
DomEnvironment options.

Add a state enableLoadFileLazily in qqmldomastcreator to know if
loadPendingDependencies() needs to be called or not. When the file is
being lazily loaded, then loadPendingDependencies() needs to be called
to get the imports correctly loaded. When the file is not (because
WithSemanticAnalysis is disabled in the DomEnvironment), then
the imports are already being processed in loadPendingDependencies() and
calling it again recursively breaks loadPendingDependencies().

Allow the qqmldomastcreator to commit lazily loaded dependencies to the
validEnv by saving a pointer in domEnvironment of the latest used
domEnvironment. And do a commitToBase after loading dependencies from
imported modules.

Task-number: QTBUG-122645
Change-Id: I3ca8bc011dcb5920fffab786cdb9e748667be48e
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>