QQuickMenu: don't set QPlatformMenu.visible
'visible' in QPA controls if the menu should be visible
in the menubar or not. In QML, 'visible' controls if the
menu should be open or not. Since we always want a menu
to be visible in the menubar (if it's inside of one)
we never want to set QPlatformMenu.visible to false.
The only way to currely hide a menu from the menubar
is to remove it.

Change-Id: I363c45b4ce18efb3b40d9284da456d7ce88a9c14
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>