Menu: reflect programmatic changes to non-native items in native items
Changing e.g. the enabled property of an Action would result in the
equivalent native menu item being enabled when the menu was re-opened,
but not if the change happened while the menu was already open.

While we're at it, we can ensure that the nice collection of
connections we've amassed don't linger by deleting
QQuickNativeMenuItems instantly, rather than using deleteLater.
This is safe to do since there aren't any external (e.g. QML)
references to them.

Task-number: QTBUG-69558
Change-Id: Ie2465a53fb87074c3e51c7b8a5da9c154f9d814f
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>