Android: Add QQuickView status handling API to QtQuickView
QQuickView has status() and statusChanged() to report its
status. Calling getProperty(), setProperty() and
addSignalHandler() is effective only if the QML object is
loaded and ready. We are adding an API and an interface
that reports this status to the user from the QtQuickView.

Task-number: QTBUG-122744
Change-Id: I8e64ba554f2c4f18308ab1411532c75d7b1110b7
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit 04b061b8b1a2028d9a2fc8910ba48d84082106af)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>