Clear tooltip transition manager state when its parent window destroyed
The tooltip transition state is not reset when its parent window is
destroyed. It happens when the tooltip parent is destroyed and the state
of the transition manager (enter transition) has not been reset (as the
window is not available). This previous state of the transition manager
doesn't allow setting the visibility of the tooltip from that point on.
This issue can be reproduced with the test case ToolTip::test_attached
(material style).

This patch reset the state of transition manager whenever visibility
of the tooltip been set.

Amend patch c16dc16cafa78ea2a3d1ef6e374f46d6632da779.

Pick-to: 6.6
Change-Id: I53719ebb6809df98fce16c68d950874f38af8fe8
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
(cherry picked from commit d03c585195d18d2400b37954c101f00e486691e2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>