QV4 garbage collection: Allocate white during gc run
Now that we properly employ the WriteBarrier, we no longer need to use
black allocations. This reduces the amount of floating garbage, and
avoids potential bugs with heap classes like QV4::MemberData where we
rely on it being marked so that its constituents don't have to be
protected during initialization – which would fail when we it is already
black, as it wouldn't be pushed on the MarkStack afterwards.

Fixes: QTBUG-121910
Change-Id: Ie5e0f9461fa5c0e4a293da3d3c978da73743e417
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>