Menu: fix an assertion failure when destroying menus in menubars
Previously recursivelyCreateNativeMenuItems was called after the call
to nativeHandle in syncWithRequestNative, and it's a requirement that
it is called when the handle is created. QQuickMenuBar calls
nativeHandle (which calls createNativeMenu) in several places, so it
makes sense to just move recursivelyCreateNativeMenuItems into
createNativeMenu.

This fixes an assertion failure when destroying menus, because the
native items were never created, so the index would never be valid.

The functions that recursivelyCreateNativeMenuItems calls assumes
the Menu is completed, and this causes an assertion failure as MenuBar
didn't wait until component completion before creating its native
handle. Fix this by returning early from syncNativeMenuBarVisible
if component completion has happened yet.

Task-number: QTBUG-69558
Change-Id: I7702b32cdf2b666f74e6ebe8f06a6d8c1c7cdb7d
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>