QQuickMenu: don't assume menuBar is the direct parent of a menu
A Menu can sometimes be grandchild of the MenuBar, rather
than being a direct child. This will e.g happen in case
of nested sub menus, or if the Menu is a child of a
MenuBarItem (which is a child of the MenuBar).

This patch will therefore add a function resolveMenuBar()
that ensures that we search for a menu bar up the
parent chain whenever we need to access the MenuBar.

Change-Id: I6cf158b642de6a3b0ad18f45ce4a488e49421f27
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>