QQuickMenu: sync native menu when the native menu is created
When a native menu is created, we need to ensure that we
sync it's properties, such as the text. This is normally
done when the component is complete, but since the
menu (or the menubar) might not be requested native
at that point, but later, it's better to do this when
we actually create the native menu.

Change-Id: Ie2b38ab6f44ca16e12cd20801049e449b01d1a61
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>