Mostly revert "Make properties in the particles module FINAL"
Making properties FINAL is an API break not covered by any QUIP rule.
However, this does not apply to the attached type whose properties were
marked FINAL in the referenced commit, as it is not possible to derive
from it anyway.

This reverts the majority of commit
cb82b9f2addf25db34970ebb48f69291901ad7ba.

Pick-to: 6.6 6.7
Change-Id: Ib001417f0b4236f1c8e068ec5f238a4b9731f8cc
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>