Introduce plugin support for qmlls and qmlls quick plugin
Introduce the plugin support for qmlls, and add an (empty) snippet qmlls
plugin that will provide quick snippets in the future.

Add the QQmlLSPlugin plugin interface, and a QmlLSQuickPlugin that
implements it. Use QFactoryLoader to load the plugins.

Move the quick snippet code from qqmllscompletion to the newly
introduced quick qmlls plugin for snippet completion.

Add a test in tst_qmlls_modules to make sure that the plugin is loaded
correctly in qmlls.

Fixes: QTBUG-119969
Change-Id: If9cfedfbf166aff0eca8fe7008b0564035cd0153
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>