Replace QSGTextNode::smooth with QSGTextNode::filtering
The "smooth" property was intended to match the one in QQuickItem,
which uses a less technical name (but also less precise). Since
QSGTextNode is scenegraph API, it makes sense to expose the actual
property of the texture instead of trying to mimic the vague name
in QQuickItem.

Change-Id: I6fb2ac637a654e1df677bf95f43c28b0bc0661a1
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 416312130ed136a2e3f54c00b71400108fb75161)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>