QQuickMenuBar: search for the root menu and menubar using QObject::parent only
For a menu inside a menu bar, qquickmenubar used to parent the
menu inside its MenuBarItem using parentItem(). But this has
changed, so currently a menu is parented into a MenuBarItem
using QObject::parent instead. This is also the same parent
chain that is used for parenting sub menus into parent menus.

This means that we can now simplify the code in QQuickMenu so that
we don't need to take three different parent chains into account
when searching for the root menu or menubar of a menu
(QObject::parent, QQuickMenu::parentItem, and QQuickItem::parentItem).
The current implementation was anyway flawed, since it would
fail to continue follow the QObject::parent for QQuickItems where
QQuickItem::parentItem return nullptr.

Change-Id: I119a2411d159032fb3fa63ed3f5c601ccc890cb7
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>