QQuickWidget: Play nice with Vulkan and offscreen grabbing
The texture does not have the UsedAsTransferSource flag set. With
Vulkan this results in a validation layer warning (if that is enabled)
when issuing the VkImage -> buffer copy during the grab. Add the flag.

We also need to ensure there is a QVulkanInstance
(VkInstance). Normally we do not care about this and just forward what
the toplevel/backingstore use, but this is not there when trying to
grab with a not-yet-shown widget.  Use the helper singleton
(QVulkanDefaultInstance) to get a QVulkanInstance. (this is also what
the backingstore and Qt Quick uses)

Pick-to: 6.7
Change-Id: I9f5b4a67a1a7862e6b89e76f2a6d55c8a2be1972
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>