Shortcut: Match window shortcut when its window is active
The window shortcut declared in the embedded QQuickView did not get
activated unless the QQuickView was the focus window, instead of getting
activated because the containing window was active.

Now, the shortcut's window is checked to be active, and if it's not,
the matcher dose not match the shortcut.

The testcase creates a shortcut event in a window which contains an
embedded window shortcut. It should pass when the embedded shortcut's
window is active.

Task-number: QTBUG-121785
Change-Id: I0a27904d73492f1f8ee129441872cc5f2482b785
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>