QQuickAttachedPropertyPropagator: fix child window propagation
After the recent changes to window code, a child window's transient
parent is not available as early as it used to be. Account for this
by listening to changes in the window's parent as we do for items.

Fixes: QTBUG-122008
Pick-to: 6.7.0
Change-Id: I2cd93e3c487e7ce4567e740886ef65a756c1350f
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit 5f7b25a694eb46cc520b6dd2651f1e3f99e310f0)
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>