Fix advance of text item with Text.RichText
This amends bf7ed10ce1ed31e24a35bc786424656308c91256 which
calculated the correct baseline position of Text.RichText.
However, this caused the advance of the text item to be wrong
if the text lines had different ascents. This accounts for this
by calculating the distance between the baselines of the text
instead of their top level positions.

Change-Id: I7259724a84ff014f3c26d5cdc49a45378cc0a91c
Reviewed-by: Lars Knoll <lars@knoll.priv.no>