Shortcut: Don't match application shortcut if its window is not visible
When a window containing a Quick shortcut was closed, we would still
match the application shortcut, which doesn't match the behavior of Qt
Widgets.

We now check the visible state of the window containing the shortcut
before matching it.

The testcase creates a shortcut event in another window. It should fail
when the window which contains embedded application shortcut is closed.

Task-number: QTBUG-121785
Change-Id: I8cf5d3797102e0a1b43a91b347bdbfdd8cfc39e2
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>