Accept mouse or touchpad in snippets with acceptedDevices
This is a workaround for platforms (such as Wayland and macOS) that are
sometimes failing to distinguish mouse and touchpad. It's better for
illustrative purposes anyway, since a laptop user may really be using a
touchpad rather than a mouse, and we hope to eventually distinguish
them on as many platforms as possible.

Task-number: QTBUG-63363
Task-number: QTBUG-112432
Pick-to: 6.6 6.5
Change-Id: Ic7a374f1257d4aa57f29385b44da85ccaf4f5ec6
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
Reviewed-by: Doris Verria <doris.verria@qt.io>
(cherry picked from commit 32a56ba8f08ddefa3d20c3003ed8599553b6d960)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>