QQuickRt: support resolving into the provided native texture
One might naively think multisample rendering in XR can be achieved
in a portable manner by setting XrSwapchainCreateInfo::sampleCount
to > 1, and then rendering with a multisample-enabled pipelines into
the multisample textures (or multisample texture array, if multiview)
that the XrSwapchain hands out.

In practice this works with 1 out of 5 OpenXR platforms tested, the
rest just claims the maxSampleCount is 1.

Therefore, to implement MSAA in Quick 3D XR, we need to follow the
classic approach of rendering into an intermediate multisample
texture/renderbuffer and then setup resolving into the OpenXR-provided
non-multisample texture at the end of the render pass.

To facilitate this and to reduce a huge duplication of logic, add
the appropriate functions to QQuickRenderTarget, and build in
the support for setting up a color attachment (and an intermediate
color buffer).

Also fix some doc typos and incorrect overload tags.

Change-Id: I5911c49b6a11d423f22a7cc8401a17b3d0310615
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>