Stabilize tst_qquicktextedit::largeTextObservesViewport
Ensure that firstBlockInViewport is set in all cases in
QQuickTextEdit::updatePaintNode() where there is a viewport and we
found one or more blocks that are within the viewport. Since
5ce225d6d310005f032ff319da8df15f7e10851a it's being used in a
conditional check, so it's no longer only for the autotest; and
there was an else case where it wouldn't get set. Perhaps that
was happening in CI. Amends f878019332cc496cc00acd9554640ca198aa10a6

Pick-to: 6.6 6.7
Fixes: QTBUG-109488
Change-Id: Ib7ccd4aee4e41ff84cf0ebaca00e295568bcb9f9
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>