WriteBarrier: Introduce Pointer and HeapObjectWrapper abstractions
We already have QV4::Pointer as a way to keep references to heap items
which need to be protected by the WriteBarrier. However, it is only meant
to be used in other HeapItems. WriteBarrier::Pointer is its complement
which can be used in the few classes which live on the C++ heap and
store references to HeapItems.
In the same vein, introduce HeapObjectWrapper, which has the same
purpose but is trivial (so that it can be used in classes that are
required to be trivial).

Change-Id: I4193fa4ae4fe7c997ff5769168ac15d60bf3726d
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>