Render gradient-filled text with paths in VectorImage
The Text component in Qt does not support gradient fills, so we
need to implement this via Qt Quick Shapes instead, basically
converting the whole text to a path and using that.

It removes the pointers to the gradient in the style resolver,
since we want to be able to have support for spans with their
own gradients so we need to copy the data out from the formats
and not risk keeping dangling pointers around.

Note: To ensure we get all the rich text processing from
QTextDocument, we get the glyphs from the layout and convert
these to shapes instead of using the PathText convenience class.

Fixes: QTBUG-122114
Change-Id: Iac66ae3dc767712395ed6cdd9a52696cd5482077
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>