Fix whitespace mode in SVG visitor
When using the RichText format, we need to make sure the
HTML parser does not compress whitespace. The preserve
whitespace mode in SVG will preserve spaces, but it will
still remove manual line breaks, so in addition we have
to remove all new lines from the input text. We do this
by duplicating the code from QSvgText::draw_helper() where
tabs and newlines are converted to spaces.

Fixes: QTBUG-122115
Change-Id: I97a8a739a4ddc5bac3ce794d515c56dd933265ec
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>