Give focus to parent window when focus chain wraps
If the focus chain wraps when tabbing (prev == last && next == first),
we need to give a chance to the parent window to gain focus so that
the focus can be passed to the next/prev object in the focus chain.
To give focus to the prev/next target, override new virtual
setFocusToTarget for QQuickWindowPrivate.

Task-number: QTBUG-121789
Change-Id: Ibe91af53ca622e7fe2b7fc662a95f1a5d7cb479b
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>