Optimize QQuickRt behavior with implicitly created buffers
Creating a new depth-stencil, or with the new auto-resolving variants,
a multisample color buffer on every setRenderTarget() is not ideal.
(e.g. Quick 3D XR at the moment uses the convenient fromVulkanImage(),
fromOpenGLTexture(), etc. constructors and then setRenderTarget(), on
every frame (multiview) - or per eye per frame (non-multiview) -
passing in the XrSwapchain-provided native image) As QQuickRenderTarget
is merely a value class referencing some resources plus contains
metadata, attempting to keep around QQuickRenderTarget instances
on the user side is not effective; from the renderer's point of view
every call to setRenderTarget() is effectively a full change, even
though we might have used that QQuickRenderTarget before.

At minimum, we should try keeping the already existing depth-stencil,
multisample, and other implicit (auto-created) buffers if they are
suitable for the new QQuickRenderTarget. (suitable = size matches,
format matches, other relevant properties match) This is implemented
in this patch.

That is then suitable to speed up the vast majority of common cases,
and will avoid creating new implicit buffers on every frame with
Quick 3D XR for example. Those who want to continuously render with
a different size is not helped by this, but it is not clear if there
is such a use case in practice to begin with. Introducing pools and
caches and such are not reasonable here, since keeping all sorts
of depth-stencil buffers with various sizes around causes
more problems than it solves. (who trims it, how much it can grow,
etc.) Power users should anyway resort to managing their own
QRhiRenderTarget objects and passing that in via
fromRhiRenderTarget(). (this is something to be considered for
Quick 3D XR as well later on; when using the API-specific
QQuickRt constructors the QRhiTextureRenderTarget is still
going to be remade after every setRenderTarget(), which is also
not entirely ideal, although the details depend on the
underlying 3D API as well)

Change-Id: I5754a4809038e9688cbd5618324383886cea7010
Reviewed-by: Andy Nichols <andy.nichols@qt.io>